wallpaper-processor/requirements.txt
numpy==2.2.4
pillow==11.1.0
requests==2.32.3
scikit_learn==1.6.1
torch==2.6.0
transformers==4.50.3
kafka-python==2.0.2
python-dotenv==1.0.0